refactor(browser-test): add a helper function for tests in browsers
